requirements.txt
accelerate==1.0.1
addict==2.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.1.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
huggingface-hub==0.27.0
idna==3.10
imageio==2.35.1
importlib_resources==6.4.5
Jinja2==3.1.5
kiwisolver==1.4.7
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mmengine==0.10.5
mpmath==1.3.0
networkx==3.1
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
packaging==24.2
pandas==2.0.3
peft==0.13.0
pillow==10.4.0
platformdirs==4.3.6
psutil==6.1.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
scikit-image==0.21.0
scipy==1.10.1
six==1.17.0
sympy==1.13.3
termcolor==2.4.0
thop==0.1.1.post2209072238
tifffile==2023.7.10
timm==1.0.12
tokenizers==0.20.3
tomli==2.2.1
torch==2.4.1
torchstat==0.0.7
torchvision==0.19.1
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yapf==0.43.0
zipp==3.20.2
